feat: Implement SmithyXmlSerializer for XML serialization and deserialization
- Added SmithyXmlSerializer class for handling XML serialization and deserialization of Smithy models.
- Introduced methods for serializing objects to XML and deserializing XML back to objects.
- Implemented support for XML traits such as xmlName and xmlFlattened.
- Created unit tests for SmithyXmlSerializer to validate serialization and deserialization functionality.
- Updated existing code to utilize the new SmithyPayloadDocuments for serialization in various places.
- Refactored CSharpShapeGenerator to accommodate changes in serialization logic.
- Added project references and necessary using directives for XML handling.
