+++ src/SmithyNet.Client/SmithyJsonPayloadCodec.cs
@@ -15,32 +15,8 @@ public sealed class SmithyJsonPayloadCodec : ISmithyPayloadCodec
         return Encoding.UTF8.GetBytes(SmithyJsonSerializer.Serialize(value));
     }
 
-    public byte[] SerializeMembers(string rootName, IReadOnlyDictionary<string, object?> members)
-    {
-        ArgumentException.ThrowIfNullOrWhiteSpace(rootName);
-        ArgumentNullException.ThrowIfNull(members);
-
-        return Serialize(members);
-    }
-
     public T Deserialize<T>(byte[] content)
     {
         return SmithyJsonSerializer.Deserialize<T>(Encoding.UTF8.GetString(content));
-    }
-
-    public T DeserializeMember<T>(byte[] content, string name)
-    {
-        ArgumentException.ThrowIfNullOrWhiteSpace(name);
-
-        var json = Encoding.UTF8.GetString(content);
-        if (string.IsNullOrWhiteSpace(json))
-        {
-            return default!;
-        }
-
-        using var document = JsonDocument.Parse(json);
-        return document.RootElement.TryGetProperty(name, out var value)
-            ? SmithyJsonSerializer.Deserialize<T>(value.GetRawText())
-            : default!;
     }
 }
