+++ src/SmithyNet.Client/SmithyNet.Client.csproj
@@ -3,5 +3,6 @@
     <ProjectReference Include="../SmithyNet.Core/SmithyNet.Core.csproj" />
     <ProjectReference Include="../SmithyNet.Http/SmithyNet.Http.csproj" />
     <ProjectReference Include="../SmithyNet.Json/SmithyNet.Json.csproj" />
+    <ProjectReference Include="../SmithyNet.Xml/SmithyNet.Xml.csproj" />
   </ItemGroup>
 </Project>
