+++ src/SmithyNet.Http/SmithyHttpResponse.cs
@@ -1,5 +1,5 @@
-using System.Text;
 using System.Net;
+using System.Text;
 
 namespace SmithyNet.Http;
