+++ src/SmithyNet.Xml/SmithyNet.Xml.csproj
@@ -1,0 +1,5 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <ItemGroup>
+    <ProjectReference Include="../SmithyNet.Core/SmithyNet.Core.csproj" />
+  </ItemGroup>
+</Project>
